// packages/ts/src/types/ast.ts
export interface AbstractKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.AbstractKeyword;
	readonly parent: ModifierParent;
}
export type AccessorDeclaration =
	| GetAccessorDeclaration
	| SetAccessorDeclaration;
export interface AccessorKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.AccessorKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface AsyncKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.AsyncKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface BigIntLiteral extends ts.Node {
	hasExtendedUnicodeEscape?: boolean;
	isUnterminated?: boolean;
	readonly kind: ts.SyntaxKind.BigIntLiteral;
	readonly parent:
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| FunctionOrConstructorTypeNodeBase
		| JSDocFunctionType
		| LeftHandSideExpressionParent
		| LiteralTypeNode
		| MethodDeclaration
		| SemicolonClassElement
		| TypeElement;
	text: string;
}
// ...
export type BindingName = BindingPattern | Identifier;
// ...
export interface ComputedPropertyName extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.ComputedPropertyName;
	readonly parent:
		| BindingElement
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| EnumMember
		| FunctionOrConstructorTypeNodeBase
		| JSDocFunctionType
		| JsxSpreadAttribute
		| MethodDeclaration
		| PropertyAssignment
		| PropertyDeclaration
		| SemicolonClassElement
		| SpreadAssignment
		| TypeElement;
}
// ...
export interface ConstKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ConstKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface DeclareKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.DeclareKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface DefaultKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.DefaultKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface ExportKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ExportKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface FunctionTypeNode extends ts.Node {
	readonly kind: ts.SyntaxKind.FunctionType;
	readonly name?: PropertyName;
	readonly parameters: ts.NodeArray<ParameterDeclaration>;
	readonly parent: TypeNodeParent;
	readonly type: TypeNode;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration> | undefined;
}
// ...
export interface Identifier extends ts.Node {
	readonly kind: ts.SyntaxKind.Identifier;
	readonly parent:
		| ArrayLiteralExpression
		| ArrowFunction
		| AsExpression
		| AwaitExpression
		| BinaryExpression
		| BindingElement
		| BreakStatement
		| CallExpression
		| CaseClause
		| ClassExpression
		| ClassStaticBlockDeclaration
		| CommaListExpression
		| ComputedPropertyName
		| ConditionalExpression
		| ConstructorDeclaration
		| ContinueStatement
		| DeclarationStatement
		| Decorator
		| DeleteExpression
		| ElementAccessExpression
		| EnumMember
		| ExportSpecifier
		| ExpressionStatement
		| ExternalModuleReference
		| FunctionExpression
		| FunctionOrConstructorTypeNodeBase
		| IfStatement
		| ImportAttribute
		| ImportClause
		| ImportDeclaration
		| ImportSpecifier
		| IterationStatement
		| JSDocFunctionType
		| JSDocLink
		| JSDocLinkCode
		| JSDocLinkPlain
		| JSDocMemberName
		| JSDocNamespaceDeclaration
		| JSDocPropertyLikeTag
		| JSDocReturnTag
		| JSDocTemplateTag
		| JsxAttributeLike
		| JsxClosingElement
		| JsxExpression
		| JsxNamespacedName
		| JsxOpeningElement
		| JsxSelfClosingElement
		| JsxTagNamePropertyAccess
		| LabeledStatement
		| MetaProperty
		| MethodDeclaration
		| NamedTupleMember
		| NamespaceDeclaration
		| NamespaceExport
		| NamespaceImport
		| NewExpression
		| NodeWithTypeArguments
		| NonNullExpression
		| ParameterDeclaration
		| ParenthesizedExpression
		| PartiallyEmittedExpression
		| PostfixUnaryExpression
		| PrefixUnaryExpression
		| PropertyAccessExpression
		| PropertyAssignment
		| PropertyDeclaration
		| QualifiedName
		| ReturnStatement
		| SatisfiesExpression
		| SemicolonClassElement
		| ShorthandPropertyAssignment
		| SpreadAssignment
		| SpreadElement
		| SwitchStatement
		| TaggedTemplateExpression
		| TemplateSpan
		| ThrowStatement
		| TypeAssertion
		| TypeElement
		| TypeOfExpression
		| TypeParameterDeclaration
		| TypePredicateNode
		| VariableDeclaration
		| VoidExpression
		| WithStatement
		| YieldExpression;

	/**
	 * Prefer to use `id.unescapedText`. (Note: This is available only in services, not internally to the TypeScript compiler.)
	 * Text of identifier, but if the identifier begins with two underscores, this will begin with three.
	 */
	readonly escapedText: ts.__String;
	readonly text: string;
}
// ...
export interface InKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.InKeyword;
	readonly parent: ModifierParent;
}
export interface InterfaceDeclaration extends ts.Node {
	readonly heritageClauses?: ts.NodeArray<HeritageClause>;
	readonly kind: ts.SyntaxKind.InterfaceDeclaration;
	readonly members: ts.NodeArray<TypeElement>;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly name: Identifier;
	readonly parent: StatementParent;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration>;
}
// ...
export interface MappedTypeNode extends ts.Node {
	readonly kind: ts.SyntaxKind.MappedType;
	readonly nameType?: TypeNode;
	readonly parent: TypeNodeParent;
	readonly questionToken?: MinusToken | PlusToken | QuestionToken;
	readonly readonlyToken?: MinusToken | PlusToken | ReadonlyKeyword;
	readonly type?: TypeNode;
	readonly typeParameter: TypeParameterDeclaration;
	/** Used only to produce grammar errors */
	readonly members?: ts.NodeArray<TypeElement>;
}
// ...
export interface MethodSignature extends ts.Node {
	readonly kind: ts.SyntaxKind.MethodSignature;
	readonly modifiers?: ts.NodeArray<Modifier>;
	readonly name: PropertyName;
	readonly parameters: ts.NodeArray<ParameterDeclaration>;
	readonly parent: InterfaceDeclaration | MappedTypeNode | TypeLiteralNode;
	readonly questionToken?: QuestionToken | undefined;
	readonly type?: TypeNode | undefined;
	readonly typeParameters?: ts.NodeArray<TypeParameterDeclaration> | undefined;
}
// ...
export type Modifier =
	| AbstractKeyword
	| AccessorKeyword
	| AsyncKeyword
	| ConstKeyword
	| DeclareKeyword
	| DefaultKeyword
	| ExportKeyword
	| InKeyword
	| OutKeyword
	| OverrideKeyword
	| PrivateKeyword
	| ProtectedKeyword
	| PublicKeyword
	| ReadonlyKeyword
	| StaticKeyword;
export type ModifierLike = Decorator | Modifier;
// ...
export interface NoSubstitutionTemplateLiteral extends ts.Node {
	hasExtendedUnicodeEscape?: boolean;
	isUnterminated?: boolean;
	readonly kind: ts.SyntaxKind.NoSubstitutionTemplateLiteral;
	readonly parent:
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| FunctionOrConstructorTypeNodeBase
		| JSDocFunctionType
		| LeftHandSideExpressionParent
		| LiteralTypeNode
		| MethodDeclaration
		| SemicolonClassElement
		| TypeElement;
	rawText?: string;
	text: string;
}
// ...
export interface OutKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.OutKeyword;
	readonly parent: ModifierParent;
}
export interface OverrideKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.OverrideKeyword;
	readonly parent: ModifierParent;
}
export interface ParameterDeclaration extends ts.Node {
	readonly dotDotDotToken?: DotDotDotToken;
	readonly initializer?: Expression;
	readonly kind: ts.SyntaxKind.Parameter;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly name: BindingName;
	readonly parent:
		| AccessorDeclaration
		| ArrowFunction
		| CallSignatureDeclaration
		| ConstructorDeclaration
		| ConstructSignatureDeclaration
		| FunctionDeclaration
		| FunctionExpression
		| FunctionOrConstructorTypeNodeBase
		| IndexSignatureDeclaration
		| JSDocFunctionType
		| MethodDeclaration
		| MethodSignature
		| SyntheticExpression;
	readonly questionToken?: QuestionToken;
	readonly type?: TypeNode;
}
// ...
export interface PrivateKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.PrivateKeyword;
	readonly parent: ModifierParent;
}
// ...
export type PropertyName =
	| BigIntLiteral
	| ComputedPropertyName
	| Identifier
	| NoSubstitutionTemplateLiteral
	| NumericLiteral
	| PrivateIdentifier
	| StringLiteral;
export interface PropertySignature extends ts.Node {
	readonly kind: ts.SyntaxKind.PropertySignature;
	readonly modifiers?: ts.NodeArray<Modifier>;
	readonly name: PropertyName;
	readonly parent: InterfaceDeclaration | MappedTypeNode | TypeLiteralNode;
	readonly questionToken?: QuestionToken;
	readonly type?: TypeNode;
}
export interface ProtectedKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ProtectedKeyword;
	readonly parent: ModifierParent;
}
export interface PublicKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.PublicKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface QuestionToken extends ts.Node {
	readonly kind: ts.SyntaxKind.QuestionToken;
	readonly parent:
		| ArrowFunction
		| ConditionalExpression
		| ConstructorDeclaration
		| FunctionDeclaration
		| FunctionExpression
		| MappedTypeNode
		| MethodDeclaration
		| ParameterDeclaration
		| PropertyDeclaration
		| TypeElement;
}
export interface ReadonlyKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ReadonlyKeyword;
	readonly parent: MappedTypeNode | ModifierParent;
}
// ...
export interface StaticKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.StaticKeyword;
	readonly parent: ModifierParent;
}
// ...
export interface TypeLiteralNode extends ts.Node {
	readonly kind: ts.SyntaxKind.TypeLiteral;
	readonly members: ts.NodeArray<TypeElement>;
	readonly parent: TypeNodeParent;
}
export type TypeNode =
	| AnyKeyword
	| ArrayTypeNode
	| BigIntKeyword
	| BooleanKeyword
	| ConditionalTypeNode
	| FunctionOrConstructorTypeNodeBase
	| IndexedAccessTypeNode
	| InferTypeNode
	| IntersectionTypeNode
	| IntrinsicKeyword
	| JSDocType
	| JSDocTypeExpression
	| LiteralTypeNode
	| MappedTypeNode
	| NamedTupleMember
	| NeverKeyword
	| NodeWithTypeArguments
	| NumberKeyword
	| ObjectKeyword
	| OptionalTypeNode
	| ParenthesizedTypeNode
	| RestTypeNode
	| StringKeyword
	| SymbolKeyword
	| TemplateLiteralTypeNode
	| TemplateLiteralTypeSpan
	| ThisTypeNode
	| TupleTypeNode
	| TypeLiteralNode
	| TypeOperatorNode
	| TypePredicateNode
	| UndefinedKeyword
	| UnionTypeNode
	| UnknownKeyword
	| VoidKeyword;
// ...
export interface TypeParameterDeclaration extends ts.Node {
	readonly kind: ts.SyntaxKind.TypeParameter;
	readonly modifiers?: ts.NodeArray<Modifier>;
	readonly name: Identifier;
	readonly parent:
		| AccessorDeclaration
		| ArrowFunction
		| CallSignatureDeclaration
		| ClassDeclaration
		| ClassExpression
		| ConstructorDeclaration
		| ConstructSignatureDeclaration
		| FunctionDeclaration
		| FunctionExpression
		| FunctionOrConstructorTypeNodeBase
		| IndexSignatureDeclaration
		| InferTypeNode
		| InterfaceDeclaration
		| JSDocFunctionType
		| JSDocTemplateTag
		| MappedTypeNode
		| MethodDeclaration
		| MethodSignature
		| TypeAliasDeclaration;
	/** Note: Consider calling `getEffectiveConstraintOfTypeParameter` */
	readonly constraint?: TypeNode;
	readonly default?: TypeNode;
	expression?: Expression;
}
// ...
export type TypeNodeParent =
	| AccessorDeclaration
	| ArrayTypeNode
	| ArrowFunction
	| AsExpression
	| CallExpression
	| CallSignatureDeclaration
	| ConditionalTypeNode
	| ConstructorDeclaration
	| ConstructSignatureDeclaration
	| FunctionDeclaration
	| FunctionExpression
	| FunctionOrConstructorTypeNodeBase
	| IndexedAccessTypeNode
	| IndexSignatureDeclaration
	| IntersectionTypeNode
	| JSDocFunctionType
	| JSDocNamepathType
	| JSDocNonNullableType
	| JSDocNullableType
	| JSDocOptionalType
	| JSDocTypeExpression
	| JSDocVariadicType
	| JsxOpeningElement
	| JsxSelfClosingElement
	| MappedTypeNode
	| MethodDeclaration
	| MethodSignature
	| NamedTupleMember
	| NewExpression
	| NodeWithTypeArguments
	| OptionalTypeNode
	| ParameterDeclaration
	| ParenthesizedTypeNode
	| PropertyDeclaration
	| PropertySignature
	| RestTypeNode
	| SatisfiesExpression
	| TaggedTemplateExpression
	| TemplateLiteralTypeSpan
	| TupleTypeNode
	| TypeAliasDeclaration
	| TypeAssertion
	| TypeOperatorNode
	| TypeParameterDeclaration
	| TypePredicateNode
	| UnionTypeNode
	| VariableDeclaration;